chore: remove native init() functions [DET-5574]
