+++ e2e_tests/tests/experiment/__init__.py
@@ -5,7 +5,6 @@ from .experiment import (
     cancel_single,
     check_if_string_present_in_trial_logs,
     create_experiment,
-    create_native_experiment,
     experiment_has_active_workload,
     experiment_state,
     experiment_trials,
@@ -14,7 +13,6 @@ from .experiment import (
     get_flat_metrics,
     get_validation_metric_from_last_step,
     maybe_create_experiment,
-    maybe_create_native_experiment,
     pause_experiment,
     root_user_home_bind_mount,
     run_basic_test,
