+++ harness/determined/experimental/__init__.py
@@ -11,6 +11,5 @@ from determined.experimental._native import (
     create,
     create_trial_instance,
     test_one_batch,
-    init_native,
-    _load_trial_on_local,
+    _load_trial_for_checkpoint_export,
 )
