--- harness/determined/experimental/estimator/__init__.py
@@ -1,1 +1,0 @@
-from determined.experimental.estimator._estimator_native import init
