--- harness/determined/experimental/keras/__init__.py
@@ -1,1 +1,0 @@
-from determined.experimental.keras._tf_keras_native import init
