+++ harness/determined/load/__init__.py
@@ -1,12 +1,10 @@
-from determined.load._load_implementation import (
+from determined.load._native import (
     RunpyGlobals,
-    load_native_implementation,
-    trial_class_from_entrypoint,
+    load_native,
 )
-from determined.load._load_trial_controller import (
-    load_native_implementation_controller,
-    load_controller_from_trial,
-    load_trial_implementation_controller,
+from determined.load._trial import (
+    load_trial,
     prepare_controller,
     prepare_tensorboard,
+    trial_class_from_entrypoint,
 )
