+++ harness/determined/load/_trial.py
@@ -1,5 +1,7 @@
+import importlib
 import logging
 import pathlib
+import sys
 from typing import Optional, Tuple, Type, cast
 
 import determined as det
@@ -7,7 +9,62 @@ from determined import horovod, load, tensorboard, workload
 from determined.common import check
 
 
-def load_controller_from_trial(
+def trial_class_from_entrypoint(entrypoint_spec: str) -> Type[det.Trial]:
+    """
+    Load and initialize a Trial class from an entrypoint specification.
+
+    An entrypoint specification is expected to take the form:
+
+        "<module>:<object reference>"
+
+    <module> specifies the module containing the trial class within the model
+    definition, relative to the root.
+
+    <object reference> specifies the naming of the trial class within the
+    module. It may be a nested object delimited by dots.
+
+    Examples:
+
+        "model_def:CIFAR10Trial": expects a "CIFAR10Trial" class that is
+        defined in a file model_def.py
+
+        "my_lib.trial:trial_classes.NestedTrial": expects a "NestedTrial"
+        class that is an attribute of `trial_classes`, where `trial_classes` is
+        defined in a file my_lib/trial.py
+
+    Note that this follows the entrypoints specification loading logic defined
+    in [1] with a single difference: the directory name of the model definition
+    is prefixed to <module>, or used as the module if <module> is empty.
+
+    [1] https://packaging.python.org/specifications/entry-points/
+    """
+
+    logging.info(f"Loading Trial implementation with entrypoint {entrypoint_spec}.")
+    module, qualname_separator, qualname = entrypoint_spec.partition(":")
+
+    # Exporting checkpoints reliably requires instantiating models from user
+    # trials and loading their weights. The user may load multiple trials into
+    # the same process. If the trials have the same module name, ie. model_def,
+    # python will only load the module once. Thus, it would be impossible to
+    # load trials from different experiments into the same process. To avoid
+    # this, we remove the module name from sys.modules if it already exists to
+    # force python to load the module regardless of its name.
+    if module in sys.modules:
+        sys.modules.pop(module)
+
+    obj = importlib.import_module(module)
+    if qualname_separator:
+        for attr in qualname.split("."):
+            obj = getattr(obj, attr)
+
+    check.check_issubclass(
+        obj, det.Trial, "Invalid type for specified 'entrypoint' ({})".format(entrypoint_spec)
+    )
+
+    return cast(Type[det.Trial], obj)
+
+
+def load_trial(
     trial_class: Type[det.Trial],
     env: det.EnvContext,
     workloads: workload.Stream,
@@ -50,75 +107,6 @@ def load_controller_from_trial(
     )
 
 
-def load_trial_implementation_controller(
-    env: det.EnvContext,
-    workloads: workload.Stream,
-    load_path: Optional[pathlib.Path],
-    rendezvous_info: det.RendezvousInfo,
-    hvd_config: horovod.HorovodContext,
-) -> det.TrialController:
-    trial_class = load.trial_class_from_entrypoint(env.experiment_config["entrypoint"])
-    return load_controller_from_trial(
-        trial_class=trial_class,
-        env=env,
-        workloads=workloads,
-        load_path=load_path,
-        rendezvous_info=rendezvous_info,
-        hvd_config=hvd_config,
-    )
-
-
-def load_native_implementation_controller(
-    env: det.EnvContext,
-    workloads: workload.Stream,
-    load_path: Optional[pathlib.Path],
-    rendezvous_info: det.RendezvousInfo,
-    hvd_config: horovod.HorovodContext,
-) -> det.TrialController:
-    check.true(
-        env.experiment_config.native_enabled(),
-        "Experiment configuration does not have an internal.native "
-        f"configuration: {env.experiment_config}",
-    )
-
-    context, trial_class, controller_class = load.load_native_implementation(
-        env, hvd_config, rendezvous_info
-    )
-
-    if trial_class is not None:
-        return load_controller_from_trial(
-            trial_class=trial_class,
-            env=env,
-            workloads=workloads,
-            load_path=load_path,
-            rendezvous_info=rendezvous_info,
-            hvd_config=hvd_config,
-        )
-
-    else:
-        # Framework-specific native implementation.
-        check.is_not_none(
-            controller_class,
-            "The class attribute `trial_controller_class` is "
-            "None; please set it the correct subclass of `det.TrialController`",
-        )
-        check.is_subclass(
-            controller_class,
-            det.TrialController,
-            "The class attribute `trial_controller_class` is "
-            "not a valid subclass of `det.TrialController`",
-        )
-        logging.info(f"Creating {controller_class.__name__} with {type(context).__name__}.")
-        return cast(det.TrialController, controller_class).from_native(
-            context=cast(det.NativeContext, context),
-            env=env,
-            workloads=workloads,
-            load_path=load_path,
-            rendezvous_info=rendezvous_info,
-            hvd_config=hvd_config,
-        )
-
-
 def prepare_controller(
     env: det.EnvContext,
     workloads: workload.Stream,
@@ -131,13 +119,10 @@ def prepare_controller(
     """
 
     if env.experiment_config.native_enabled():
-        controller = load_native_implementation_controller(
-            env, workloads, load_path, rendezvous_info, hvd_config
-        )
+        controller = load.load_native(env, workloads, load_path, rendezvous_info, hvd_config)
     else:
-        controller = load_trial_implementation_controller(
-            env, workloads, load_path, rendezvous_info, hvd_config
-        )
+        trial_class = trial_class_from_entrypoint(env.experiment_config["entrypoint"])
+        controller = load_trial(trial_class, env, workloads, load_path, rendezvous_info, hvd_config)
 
     return controller
